ECOM CHANGES
ECOM APP CHANGES DUMPED

+++ customer_service-reactive/src/main/java/com/shoppix/customer_service_reactive/model/Cart.java
@@ -14,7 +14,7 @@ public class Cart {
 
     public List<CartProduct> cartProducts;
 
-    public int totalPrice;
+    public double totalPrice;
 
     public String eventStatus;
